adding additional checks

+++ init.sh
@@ -8,6 +8,10 @@ make_symlink () {
 	if [ ! -d $(dirname $dest_path) ]; then
 		mkdir -p $(dirname $dest_path)
 	fi
+	if [[ -f $dest_path || -d $dest_path ]]; then
+		echo "$dest_path already exists. Skipping symlink creation"
+		return 0
+	fi
 
 	echo "Making symlink from $orig_path to $dest_path"
 	ln -s $orig_path $dest_path
@@ -15,3 +19,4 @@ make_symlink () {
 
 # Create symlinks to config files
 make_symlink $DOTFILE_DIR/.ssh/config $HOME/.ssh/config
+make_symlink $DOTFILE_DIR/.config/sway $HOME/.config/sway
